feat: make a lot of QOL changes and prep for easy local development
