+++ frontend/.env.template
@@ -1,0 +1,3 @@
+NEXT_PUBLIC_MAPBOX_ACCESS_TOKEN=
+NEXT_PUBLIC_MAPBOX_STYLE_URL=
+API_URL=
