+++ frontend/components/Map.tsx
@@ -5,150 +5,148 @@ import mapboxgl from "mapbox-gl";
 import "mapbox-gl/dist/mapbox-gl.css";
 
 interface dataFormat {
-    building: string;
-    building_code: string;
-    building_status: string;
-    rooms: {
-        [key: string]: {
-            roomNumber: string;
-            slots: { StartTime: string; EndTime: string; Status: string }[];
-        };
+  building: string;
+  building_code: string;
+  building_status: string;
+  rooms: {
+    [key: string]: {
+      roomNumber: string;
+      slots: { StartTime: string; EndTime: string; Status: string }[];
     };
-    coords: [number, number];
-    distance: number;
+  };
+  coords: [number, number];
+  distance: number;
 }
 
 export default function Map({
-    data,
-    handleMarkerClick,
-    userPos,
+  data,
+  handleMarkerClick,
+  userPos,
 }: {
-    data: dataFormat[];
-    handleMarkerClick: (building: string) => void;
-    userPos: any;
+  data: dataFormat[];
+  handleMarkerClick: (building: string) => void;
+  userPos: any;
 }) {
-    const mapRef = useRef<mapboxgl.Map | null>(null);
-    const mapContainerRef = useRef<HTMLDivElement | null>(null);
+  const mapRef = useRef<mapboxgl.Map | null>(null);
+  const mapContainerRef = useRef<HTMLDivElement | null>(null);
 
-    const [center, setCenter] = useState<[number, number]>([-80.5425, 43.4695]);
-    const [zoom, setZoom] = useState(16.25);
-    const [pitch, setPitch] = useState(52);
+  const [center, setCenter] = useState<[number, number]>([-80.5425, 43.4695]);
+  const [zoom, setZoom] = useState(16.25);
+  const [pitch, setPitch] = useState(52);
 
-    const mapboxToken = process.env.NEXT_PUBLIC_MAPBOX_ACCESS_TOKEN;
+  const mapboxToken = process.env.NEXT_PUBLIC_MAPBOX_ACCESS_TOKEN;
 
-    function getColorByStatus(status: string) {
-        switch (status) {
-            case "available":
-                return "h-2 w-2 rounded-full bg-green-400 shadow-[0px_0px_4px_2px_rgba(34,197,94,0.7)]";
-            case "unavailable":
-                return "h-2 w-2 rounded-full bg-red-400 shadow-[0px_0px_4px_2px_rgba(239,68,68,0.9)]";
-            case "upcoming":
-                return "h-2 w-2 rounded-full bg-amber-400 shadow-[0px_0px_4px_2px_rgba(245,158,11,0.9)]";
-            default:
-                return "gray";
-        }
+  const MAPBOX_STYLE_URL =
+    process.env.NEXT_PUBLIC_MAPBOX_STYLE_URL ||
+    "mapbox://styles/notakki/cm1o3v5kr00bl01pd2k7tho6i";
+
+  function getColorByStatus(status: string) {
+    switch (status) {
+      case "available":
+        return "h-2 w-2 rounded-full bg-green-400 shadow-[0px_0px_4px_2px_rgba(34,197,94,0.7)]";
+      case "unavailable":
+        return "h-2 w-2 rounded-full bg-red-400 shadow-[0px_0px_4px_2px_rgba(239,68,68,0.9)]";
+      case "upcoming":
+        return "h-2 w-2 rounded-full bg-amber-400 shadow-[0px_0px_4px_2px_rgba(245,158,11,0.9)]";
+      default:
+        return "gray";
+    }
+  }
+
+  useEffect(() => {
+    if (mapboxToken) {
+      mapboxgl.accessToken = mapboxToken;
+    } else {
+      console.error("Mapbox token is not defined");
+    }
+    mapboxToken;
+    mapRef.current = new mapboxgl.Map({
+      style: MAPBOX_STYLE_URL,
+      container: mapContainerRef.current as HTMLElement,
+      center: center,
+      zoom: zoom,
+      pitch: pitch,
+    });
+
+    mapRef.current.on("move", () => {
+      if (mapRef.current) {
+        const mapCenter = mapRef.current.getCenter();
+        const mapZoom = mapRef.current.getZoom();
+        const mapPitch = mapRef.current.getPitch();
+
+        setCenter([mapCenter.lng, mapCenter.lat]);
+        setZoom(mapZoom);
+        setPitch(mapPitch);
+      }
+    });
+
+    data.map((data) => {
+      const el = document.createElement("div");
+      el.className = getColorByStatus(data.building_status);
+
+      el.addEventListener("click", () => {
+        const accordionItem = document.getElementById(data.building_code);
+
+        setTimeout(() => {
+          if (accordionItem) {
+            accordionItem.scrollIntoView({
+              behavior: "smooth",
+              block: "start",
+            });
+          }
+        }, 300);
+
+        handleMarkerClick(data.building_code);
+      });
+
+      if (mapRef.current && data.coords) {
+        new mapboxgl.Marker(el)
+          .setLngLat([data.coords[0], data.coords[1]])
+          .addTo(mapRef.current);
+      }
+    });
+
+    if (userPos) {
+      const e2 = document.createElement("div");
+      e2.className =
+        "h-3 w-3 border-[1.5px] border-zinc-50 rounded-full bg-blue-400 shadow-[0px_0px_4px_2px_rgba(14,165,233,1)]";
+
+      new mapboxgl.Marker(e2)
+        .setLngLat([userPos[1], userPos[0]])
+        // .setLngLat([-80.5425, 43.4695])
+        .addTo(mapRef.current);
     }
 
-    useEffect(() => {
-        if (mapboxToken) {
-            mapboxgl.accessToken = mapboxToken;
-        } else {
-            console.error("Mapbox token is not defined");
-        }
-        mapboxToken;
-        mapRef.current = new mapboxgl.Map({
-            style: "mapbox://styles/notakki/cm1o3v5kr00bl01pd2k7tho6i",
-            container: mapContainerRef.current as HTMLElement,
-            center: center,
-            zoom: zoom,
-            pitch: pitch,
-        });
+    return () => {
+      if (mapRef.current) {
+        mapRef.current.remove();
+      }
+    };
+  }, []);
 
-        mapRef.current.on("move", () => {
-            if (mapRef.current) {
-                const mapCenter = mapRef.current.getCenter();
-                const mapZoom = mapRef.current.getZoom();
-                const mapPitch = mapRef.current.getPitch();
-
-                setCenter([mapCenter.lng, mapCenter.lat]);
-                setZoom(mapZoom);
-                setPitch(mapPitch);
-            }
-        });
-
-        data.map((data) => {
-            const el = document.createElement("div");
-            el.className = getColorByStatus(data.building_status);
-
-            el.addEventListener("click", () => {
-                const accordionItem = document.getElementById(
-                    data.building_code
-                );
-
-                setTimeout(() => {
-                    if (accordionItem) {
-                        accordionItem.scrollIntoView({
-                            behavior: "smooth",
-                            block: "start",
-                        });
-                    }
-                }, 300);
-
-                handleMarkerClick(data.building_code);
-            });
-
-            if (mapRef.current && data.coords) {
-                new mapboxgl.Marker(el)
-                    .setLngLat([data.coords[0], data.coords[1]])
-                    .addTo(mapRef.current);
-            }
-        });
-
-        if (userPos) {
-            const e2 = document.createElement("div");
-            e2.className =
-                "h-3 w-3 border-[1.5px] border-zinc-50 rounded-full bg-blue-400 shadow-[0px_0px_4px_2px_rgba(14,165,233,1)]";
-
-            new mapboxgl.Marker(e2)
-                .setLngLat([userPos[1], userPos[0]])
-                // .setLngLat([-80.5425, 43.4695])
-                .addTo(mapRef.current);
-        }
-
-        return () => {
-            if (mapRef.current) {
-                mapRef.current.remove();
-            }
-        };
-    }, []);
-
-    return (
-        <div className="h-[60vh] sm:w-full sm:h-full relative bg-red-500/0 rounded-[20px] p-2 sm:p-0">
-            <div
-                id="map-container"
-                ref={mapContainerRef}
-                className="opacity-100"
-            />
-            <div className="bg-[#18181b]/90 absolute bottom-10 left-2 sm:bottom-8 sm:left-0 flex flex-col gap-2 m-1 py-2.5 p-2 rounded-[16px]">
-                <div className="flex items-center gap-0">
-                    <div className="h-2 w-2 rounded-full bg-red-400 flex-none"></div>
-                    <div className="ml-2 rounded-lg px-2 py-1 text-sm w-full bg-red-700/30 text-red-300/90">
-                        unavailable
-                    </div>
-                </div>
-                <div className="flex items-center gap-0">
-                    <div className="h-2 w-2 rounded-full bg-amber-400 flex-none"></div>
-                    <div className="ml-2 rounded-lg px-2 py-1 text-sm w-full bg-amber-800/30 text-amber-300/90">
-                        opening soon
-                    </div>
-                </div>
-                <div className="flex items-center gap-0">
-                    <div className="h-2 w-2 rounded-full bg-green-400 flex-none"></div>
-                    <div className="ml-2 rounded-lg px-2 py-1 text-sm w-full bg-green-800/30 text-green-300/90">
-                        open now
-                    </div>
-                </div>
-            </div>
+  return (
+    <div className="h-[60vh] sm:w-full sm:h-full relative bg-red-500/0 rounded-[20px] p-2 sm:p-0">
+      <div id="map-container" ref={mapContainerRef} className="opacity-100" />
+      <div className="bg-[#18181b]/90 absolute bottom-10 left-2 sm:bottom-8 sm:left-0 flex flex-col gap-2 m-1 py-2.5 p-2 rounded-[16px]">
+        <div className="flex items-center gap-0">
+          <div className="h-2 w-2 rounded-full bg-red-400 flex-none"></div>
+          <div className="ml-2 rounded-lg px-2 py-1 text-sm w-full bg-red-700/30 text-red-300/90">
+            unavailable
+          </div>
         </div>
-    );
+        <div className="flex items-center gap-0">
+          <div className="h-2 w-2 rounded-full bg-amber-400 flex-none"></div>
+          <div className="ml-2 rounded-lg px-2 py-1 text-sm w-full bg-amber-800/30 text-amber-300/90">
+            opening soon
+          </div>
+        </div>
+        <div className="flex items-center gap-0">
+          <div className="h-2 w-2 rounded-full bg-green-400 flex-none"></div>
+          <div className="ml-2 rounded-lg px-2 py-1 text-sm w-full bg-green-800/30 text-green-300/90">
+            open now
+          </div>
+        </div>
+      </div>
+    </div>
+  );
 }
